api/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
python-multipart==0.0.6
joblib==1.3.2
tensorflow-cpu==2.13.0
numpy==1.24.3